feat: add permissions to middleware

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Enums\PermissionEnum;
 use App\Enums\RoleEnum;
 use App\Http\Controllers\Administrator\AdminController;
 use App\Http\Controllers\Administrator\AdminCustomerController;
@@ -70,18 +71,22 @@ Route::prefix('admin')->middleware(['auth', 'verified', "role:$admins"])->group(
 
     Route::prefix('users')->group(function () {
         Route::get('/', [AdminUserController::class, 'index'])->name('admin.users');
-        Route::get('/{user}', [AdminUserController::class, 'userShow'])->withTrashed()->name('admin.user.show');
-        Route::put('/{user}', [AdminUserController::class, 'userUpdate'])->withTrashed()->name('admin.user.update');
-        Route::delete('/{user}', [AdminUserController::class, 'userDestroy'])->name('admin.user.destroy');
-        Route::put('/{user}/password', [AdminUserController::class, 'userUpdatePassword'])->withTrashed()->name('admin.user.update.password');
-        Route::put('/{user}/restore', [AdminUserController::class, 'userRestore'])->withTrashed()->name('admin.user.restore');
-        Route::delete('/{user}/force-delete', [AdminUserController::class, 'userForceDelete'])->withTrashed()->name('admin.user.force-delete');
+        Route::middleware(['permission:'.PermissionEnum::UPDATE->value])->get('/{user}', [AdminUserController::class, 'userShow'])->withTrashed()->name('admin.user.show');
+        Route::middleware(['permission:'.PermissionEnum::UPDATE->value])->put('/{user}', [AdminUserController::class, 'userUpdate'])->withTrashed()->name('admin.user.update');
+        Route::middleware(['permission:'.PermissionEnum::UPDATE->value])->put('/{user}/password', [AdminUserController::class, 'userUpdatePassword'])->withTrashed()->name(
+            'admin.user.update.password'
+        );
+        Route::middleware(['permission:'.PermissionEnum::DELETE->value])->delete('/{user}', [AdminUserController::class, 'userDestroy'])->name('admin.user.destroy');
+        Route::middleware(['permission:'.PermissionEnum::DELETE->value])->put('/{user}/restore', [AdminUserController::class, 'userRestore'])->withTrashed()->name('admin.user.restore');
+        Route::middleware(['permission:'.PermissionEnum::DELETE->value])->delete('/{user}/force-delete', [AdminUserController::class, 'userForceDelete'])->withTrashed()->name(
+            'admin.user.force-delete'
+        );
     });
 
     Route::prefix('customers')->group(function () {
         Route::get('/', [AdminCustomerController::class, 'index'])->name('admin.customers');
-        Route::get('/{user}', [AdminCustomerController::class, 'customerShow'])->withTrashed()->name('admin.customer.show');
-        Route::put('/{user}', [AdminCustomerController::class, 'customerUpdate'])->withTrashed()->name('admin.customer.update');
+        Route::middleware(['permission:'.PermissionEnum::UPDATE->value])->get('/{user}', [AdminCustomerController::class, 'customerShow'])->withTrashed()->name('admin.customer.show');
+        Route::middleware(['permission:'.PermissionEnum::UPDATE->value])->put('/{user}', [AdminCustomerController::class, 'customerUpdate'])->withTrashed()->name('admin.customer.update');
     });
 
     Route::prefix('products/exports')->group(function () {
@@ -91,21 +96,23 @@ Route::prefix('admin')->middleware(['auth', 'verified', "role:$admins"])->group(
     });
 
     Route::prefix('products/imports')->group(function () {
-        Route::post('/', [AdminImportController::class, 'import'])->name('admin.products.import');
-        Route::get('/check', [AdminImportController::class, 'checkImport'])->name('admin.products.import.check');
+        Route::middleware(['permission:'.PermissionEnum::CREATE->value])->post('/', [AdminImportController::class, 'import'])->name('admin.products.import');
+        Route::middleware(['permission:'.PermissionEnum::CREATE->value])->get('/check', [AdminImportController::class, 'checkImport'])->name('admin.products.import.check');
     });
 
     Route::prefix('products')->group(function () {
         Route::get('/', [AdminProductController::class, 'index'])->name('admin.products');
         Route::get('/categories', [AdminProductController::class, 'searchCategories'])->name('admin.categories.search');
         Route::get('/brands', [AdminProductController::class, 'searchBrands'])->name('admin.brands.search');
-        Route::middleware(['permission:create'])->get('/create', [AdminProductController::class, 'create'])->name('admin.products.create');
-        Route::middleware(['permission:create'])->post('/create', [AdminProductController::class, 'store'])->name('admin.products.create');
-        Route::get('/{product}', [AdminProductController::class, 'productShow'])->withTrashed()->name('admin.products.show');
-        Route::post('/{product}', [AdminProductController::class, 'update'])->name('admin.products.update');
-        Route::delete('/{product}', [AdminProductController::class, 'destroy'])->name('admin.products.destroy');
-        Route::put('/{product}/restore', [AdminProductController::class, 'restore'])->withTrashed()->name('admin.products.restore');
-        Route::delete('/{product}/force-delete', [AdminProductController::class, 'forceDelete'])->withTrashed()->name('admin.products.force-delete');
+        Route::middleware(['permission:'.PermissionEnum::CREATE->value])->get('/create', [AdminProductController::class, 'create'])->name('admin.products.create');
+        Route::middleware(['permission:'.PermissionEnum::CREATE->value])->post('/create', [AdminProductController::class, 'store'])->name('admin.products.create');
+        Route::middleware(['permission:'.PermissionEnum::UPDATE->value])->get('/{product}', [AdminProductController::class, 'productShow'])->withTrashed()->name('admin.products.show');
+        Route::middleware(['permission:'.PermissionEnum::UPDATE->value])->post('/{product}', [AdminProductController::class, 'update'])->name('admin.products.update');
+        Route::middleware(['permission:'.PermissionEnum::DELETE->value])->delete('/{product}', [AdminProductController::class, 'destroy'])->name('admin.products.destroy');
+        Route::middleware(['permission:'.PermissionEnum::DELETE->value])->put('/{product}/restore', [AdminProductController::class, 'restore'])->withTrashed()->name('admin.products.restore');
+        Route::middleware(['permission:'.PermissionEnum::DELETE->value])->delete('/{product}/force-delete', [AdminProductController::class, 'forceDelete'])->withTrashed()->name(
+            'admin.products.force-delete'
+        );
     });
 });
